Add support for Minix 3.1.2a
This commit adds support to minix 3.1.2a by removing SO_REUSEADDR's
setsockopt call, which is unsupported in this version, and fixes the
LISTEN mode that was broken in this version. It also removes the
BSD-Makefile dependency and uses a default GNU Makefile instead.

This version still do not work under Linux, only in Minix.

Signed-off-by: Giuliano Belinassi <giuliano.belinassi@usp.br>